LL(1) Scripture Grammar is impossible
I have realized that LL(1) scripture grammar parsing is impossible. This is because of situations such as the following:

John 3:19; 2 Timothy 8:3.

Without two tokens of lookahead, it is impossible to see whether the 2 belongs to John as a scripture or to 2 Timothy.

However, a reverse direction parser might be possibly LL(1).